grub-switch-to-blscfg: Use install to copy GRUB binary, modules and config
By default the cp command truncates the destination before copying from the
source, so if interrupted it can lead to a file that's half written.

This behavior can be modified using the --remove-destination option, but is
usually a better choice to use the install tool for this. So let's do that.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>