10_linux.in: Enable BLS configuration if new-kernel-pkg isn't present
Currently if the the GRUB_ENABLE_BLSCFG option in /etc/default/grub hasn't
been set, the 10_linux script will generate a GRUB configuration that does
not include the blscfg command to populate the menu entries from BLS files.

But on kernel install the /usr/lib/kernel/install.d/20-grub.install script
will only add menuentry commands to the GRUB config file if the old grubby
tool and new-kernel-pkg script are installed.

So a configuration with the GRUB_ENABLE_BLSCFG option will lead to a setup
where new kernel entries are not added. Make a BLS config the default if
that option wasn't set and the new-kernel-pkg script is not present.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>