10_linux.in: Check if petitboot sysfs has a valid version
The script assumes that the petitboot sysfs entry contains a valid version
number, but this is not always the case. For example, on a Talos II system
a user reported that contained the following value:

$ cat /sys/firmware/devicetree/base/ibm,firmware-versions/petitboot
0ed84c0-p94177c1

This lead to the script wrongly trying to compare these hashes as if they
were integers values, which caused the following error when re-generating
the GRUB configuration file:

/etc/grub.d/10_linux: line 234: test: 0ed84c0-p94177c1: integer expression expected
/etc/grub.d/10_linux: line 235: test: 0ed84c0-p94177c1: integer expression expected

Check that the major and minor values are integers before attempting to do
the comparison. If these aren't numbers, then generate the menu with a set
of menuentry commands, since there's no way to know the petitboot version.

Resolves: rhbz#1921479

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>